requirements.txt
py-clob-client>=0.18.0
python-telegram-bot>=20.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
httpx>=0.25.0
